+++ projects/ngx-mask-lib/src/test/utils/test-component.component.ts
@@ -7,7 +7,7 @@ import { IConfig } from '../../lib/config';
 	selector: 'mask-test-mask',
 	template: `
 		<input
-			id="maska"
+			id="mask"
 			[mask]="mask"
 			[clearIfNotMatch]="clearIfNotMatch"
 			[dropSpecialCharacters]="dropSpecialCharacters"
@@ -24,6 +24,7 @@ import { IConfig } from '../../lib/config';
 			[hiddenInput]="hiddenInput"
 			[allowNegativeNumbers]="allowNegativeNumbers"
 			[leadZeroDateTime]="leadZeroDateTime"
+			[triggerOnMaskChange]="triggerOnMaskChange"
 		/>
 	`,
 })
@@ -59,4 +60,6 @@ export class TestMaskComponent {
 	public allowNegativeNumbers: IConfig['allowNegativeNumbers'] = false;
 
 	public leadZeroDateTime: IConfig['leadZeroDateTime'] = false;
+
+	public triggerOnMaskChange: IConfig['triggerOnMaskChange'] = false;
 }